Move all util based functions outside of elasticnet class
